+++ common/flatpak-exports.h
@@ -24,7 +24,13 @@
 #include "libglnx/libglnx.h"
 #include "flatpak-utils.h"
 #include "flatpak-bwrap.h"
-#include "flatpak-context.h"
+
+/* In numerical order of more privs */
+typedef enum {
+  FLATPAK_FILESYSTEM_MODE_READ_ONLY    = 1,
+  FLATPAK_FILESYSTEM_MODE_READ_WRITE   = 2,
+  FLATPAK_FILESYSTEM_MODE_CREATE       = 3,
+} FlatpakFilesystemMode;
 
 typedef struct _FlatpakExports FlatpakExports;
 
@@ -32,12 +38,6 @@ void flatpak_exports_free (FlatpakExports *exports);
 FlatpakExports *flatpak_exports_new (void);
 void flatpak_exports_append_bwrap_args (FlatpakExports *exports,
                                         FlatpakBwrap *bwrap);
-void flatpak_export_paths_export_context (FlatpakContext *context,
-                                          FlatpakExports *exports,
-                                          GFile *app_id_dir,
-                                          gboolean do_create,
-                                          GString *xdg_dirs_conf,
-                                          gboolean *home_access_out);
 void flatpak_exports_add_home_expose (FlatpakExports *exports,
                                       FlatpakFilesystemMode mode);
 void flatpak_exports_add_path_expose (FlatpakExports *exports,
@@ -53,8 +53,6 @@ void flatpak_exports_add_path_dir (FlatpakExports *exports,
 
 gboolean flatpak_exports_path_is_visible (FlatpakExports *exports,
                                           const char *path);
-FlatpakExports *flatpak_exports_from_context (FlatpakContext *context,
-                                              const char *app_id);
 
 G_DEFINE_AUTOPTR_CLEANUP_FUNC (FlatpakExports, flatpak_exports_free);
